requirements.txt
entsoe-py
pandas
polars
python-dotenv
matplotlib
streamlit